@@ -1,28 +1,43 @@
 import unittest
-import os
+from pathlib import Path
+from tempfile import TemporaryDirectory
+import importlib.util
+
 from autoreqgen import formatter
 
-TEST_FILE = "examples/sample_project1/unformatted.py"
+SRC = "def    foo ():\n    print('hello')\n"
+
+def has_module(name: str) -> bool:
+    return importlib.util.find_spec(name) is not None
 
 class TestFormatter(unittest.TestCase):
-
     def setUp(self):
-        os.makedirs(os.path.dirname(TEST_FILE), exist_ok=True)
-        with open(TEST_FILE, "w") as f:
-            f.write("def    foo ():\n    print('hello')")
-
-    def test_black_formatting(self):
-        # Run formatter
-        formatter.run_formatter("black", os.path.dirname(TEST_FILE))
-
-        # Check if the formatting applied (black adds double quotes)
-        with open(TEST_FILE, "r") as f:
-            content = f.read()
-            self.assertIn('print("hello")', content)
+        self.tmp = TemporaryDirectory()
+        self.tmp_path = Path(self.tmp.name)
+        self.project = self.tmp_path / "examples" / "sample_project1"
+        self.project.mkdir(parents=True, exist_ok=True)
+        self.file = self.project / "unformatted.py"
+        self.file.write_text(SRC, encoding="utf-8")
 
     def tearDown(self):
-        if os.path.exists(TEST_FILE):
-            os.remove(TEST_FILE)
+        self.tmp.cleanup()
 
-if __name__ == "__main__":
-    unittest.main()
+    @unittest.skipUnless(has_module("black"), "black not installed")
+    def test_black_formatting(self):
+        # 1) run formatter
+        formatter.run_formatter("black", str(self.project))
+
+        # 2) verify content formatted (black prefers double quotes, fixes spacing)
+        text = self.file.read_text(encoding="utf-8")
+        self.assertIn('print("hello")', text)
+        self.assertIn("def foo():", text)
+
+        # 3) idempotence: running again should keep same content
+        before = text
+        formatter.run_formatter("black", str(self.project))
+        after = self.file.read_text(encoding="utf-8")
+        self.assertEqual(before, after)
+
+    def test_unsupported_formatter_raises(self):
+        with self.assertRaises(ValueError):
+            formatter.run_formatter("not-a-tool", str(self.project))
